+++ src/pessoa.hpp
@@ -6,14 +6,14 @@
 class Pessoa {//definição da classe Pessoa
 private:
     unsigned int codigo;
-    unsigned int telefone;
+    double telefone;
     std::string nome;
 public:
-    Pessoa(unsigned int codigo, unsigned int telefone, std::string nome);//construtor da classe Pessoa
+    Pessoa(unsigned int codigo, double telefone, std::string nome);//construtor da classe Pessoa
     virtual ~Pessoa();//destrutor virtual da classe Pessoa
 
     const unsigned int getCodigo()   const;
-    const unsigned int getTelefone() const;
+    const double       getTelefone() const;
     const std::string  getNome()     const;
     //constantes que retornam seus atributos somente para leitura
 };
